requirements.txt
pillow==10.2.0
requests==2.31.0
selenium==4.17.2